+++ .storybook/preview.js
@@ -7,11 +7,15 @@ import theme from '../src/theme'
 export const parameters = {
 	actions: { argTypesRegex: '^on[A-Z].*' },
 	options: {
-		/** Sort alphabetically */
-		storySort: (a, b) =>
-			a[1].kind === b[1].kind
-				? 0
-				: a[1].id.localeCompare(b[1].id, undefined, { numeric: true })
+		storySort: {
+			method: 'alphabetical',
+			order: [
+				'Foundation',
+				['Design Tokens', 'Colors', 'Spacing'],
+				'Components',
+				'Demo'
+			]
+		}
 	},
 	designToken: {
 		disable: true