+++ src/components/Dropdown/DropdownStyles.tsx
@@ -6,9 +6,9 @@ interface StyledPopoverProps {
 
 export const StyledPopover = styled.div<StyledPopoverProps>`
 	box-shadow: 2px 4px 21px rgba(0, 0, 0, 0.28);
-	border-radius: ${({ theme }): string => theme.tokens.ref.borderRadius.md};
+	border-radius: ${({ theme }): string => theme.tokens.dropdown.borderRadius};
 	transition: all
-		${({ theme }): string => theme.tokens.ref.transition.duration.base};
+		${({ theme }): string => theme.tokens.duration.transition.base};
 	opacity: ${({ visible }): number => (visible ? 1 : 0)};
 	transform: translateY(${({ visible }): number => (visible ? 0 : -5)}px);
 	background: #fff;