+++ src/components/Menu/MenuStyles.tsx
@@ -8,8 +8,6 @@ import { Icon } from '../Icon'
 import { MenuItemIconProps } from './MenuItemIcon'
 
 export const StyledMenu = styled.ul`
-	background: #fff;
-	border-radius: 3px;
 	list-style: none;
 	margin: 0;
 	min-width: 180px;
@@ -44,7 +42,7 @@ export const StyledMenuItemAnchor = styled.a<StyledMenuItemAnchorProps>`
 			: theme.tokens.color.text.secondary};
 	user-select: none;
 	line-height: 20px;
-	border-radius: ${({ theme }): string => theme.tokens.ref.borderRadius.sm};
+	border-radius: ${({ theme }): string => theme.tokens.menu.borderRadius.item};
 	cursor: ${({ disabled }): string => (disabled ? 'default' : 'pointer')};
 
 	${({ theme, disabled }): FlattenSimpleInterpolation | undefined =>