+++ src/tokens/helpers.tsx
@@ -23,7 +23,30 @@ export const getTextColor =
 		return $color
 	}
 
-export const getUnitlessNumber = (px: string): number => parseInt(px, 10)
+export const getUnitlessNumber = (number: string): number =>
+	Number.isInteger(number) ? parseInt(number, 10) : parseFloat(number)
 
 export const getLineHeightUnitless = (lineHeightPercentage: string): number =>
 	getUnitlessNumber(lineHeightPercentage) / 100
+
+interface BoxShadowObj {
+	x: string
+	y: string
+	blur: string
+	spread: string
+	color: string
+	type: string
+}
+
+export const getBoxShadow = (shadow: BoxShadowObj | BoxShadowObj[]): string => {
+	const shadowArr = Array.isArray(shadow) ? shadow : [shadow]
+
+	return shadowArr
+		.map(
+			(shadow) =>
+				`${shadow.type === 'innerShadow' ? 'inset' : ''} ${shadow.x} ${
+					shadow.y
+				} ${shadow.blur} ${shadow.spread} ${shadow.color}`
+		)
+		.join(', ')
+}